Merged PR 10439: CMW-9130: Add policy API client and fetch policy data
## Description

* Adds enterprise API policy client leveraging the OpenAPI spec to generate policy.gen.go
* Adds mock kafka message to allow for end-to-end testing with a message that includes a policy number
* Starting to break up the models files in an attempt to stay organized
* Update README with some more set up instructions and getting an enterprise API auth token

The auth token set up is temporary until we can coordinate with CIAM team on the right strategy to auth the app. They don't have a standard process for setting up auth for new services right now.

<!--- Describe your changes in detail -->

## Motivation and Context

This leverages the OpenAPI spec from our enterprise API to add a new policy API client to fetch data. I've also made some tweaks to mock out a kafka event so we can do an end-to-end test when a message is received with a valid policy number.

<!--- Why is this change required? What problem does it solve? -->

[CMW-ticketnumber](https://se2sordev.atlassian.net/browse/CMW-ticketnumber)

## Type of Change:

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix
- [x] New feature
- [ ] Cleanup, Refactor, Update