Merged PR 10693: CMW-10251: Business rules config
## Description

Sets up the ability to pull a list of rules from a config file and make a request to the rules engine to execute each rule with an appropriate payload - inputs with policy data. Changes also include:

* Adding a logger and context to the composer and rule executor
* Adding structure to RuleOuput so expected values conform to Valid: true | false

<!--- Describe your changes in detail -->

## Motivation and Context

We need to track the list of relevant rules to execute when validating policy data.

<!--- Why is this change required? What problem does it solve? -->

[CMW-10251](https://se2sordev.atlassian.net/browse/CMW-10251)

## Type of Change:

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix
- [x] New feature
- [ ] Cleanup, Refactor, Update